Move transaction filter files that were not in the right directory
